[2628] Improve the API for custom nodes
Bug: https://github.com/eclipse-sirius/sirius-web/issues/2628
