+++ packages/diagrams/frontend/sirius-components-diagrams-reactflow/src/converter/ListNodeConverter.ts
@@ -20,10 +20,9 @@ import {
   GQLRectangularNodeStyle,
   GQLViewModifier,
 } from '../graphql/subscription/nodeFragment.types';
-import { BorderNodePosition } from '../renderer/DiagramRenderer.types';
+import { BorderNodePosition, NodeData } from '../renderer/DiagramRenderer.types';
 import { ConnectionHandle } from '../renderer/handles/ConnectionHandles.types';
-import { FreeFormNodeData } from '../renderer/node/FreeFormNode.types';
-import { IConvertEngine, INodeConverterHandler } from './ConvertEngine.types';
+import { IConvertEngine, INodeConverter } from './ConvertEngine.types';
 import { convertLineStyle } from './convertDiagram';
 import { AlignmentMap } from './convertDiagram.types';
 import { convertHandles } from './convertHandles';
@@ -31,22 +30,22 @@ import { convertLabelStyle, convertOutsideLabels } from './convertLabel';
 
 const defaultPosition: XYPosition = { x: 0, y: 0 };
 
-const toRectangularNode = (
+const toListNode = (
   gqlDiagram: GQLDiagram,
   gqlNode: GQLNode<GQLRectangularNodeStyle>,
   gqlParentNode: GQLNode<GQLNodeStyle> | null,
-  nodeDescription: GQLNodeDescription | undefined,
+  nodeDescription: GQLNodeDescription,
   isBorderNode: boolean,
   gqlEdges: GQLEdge[]
-): Node<FreeFormNodeData> => {
+): Node<NodeData> => {
   const {
     targetObjectId,
     targetObjectLabel,
     targetObjectKind,
     descriptionId,
-    id,
     insideLabel,
     outsideLabels,
+    id,
     state,
     style,
     labelEditable,
@@ -55,17 +54,22 @@ const toRectangularNode = (
   const connectionHandles: ConnectionHandle[] = convertHandles(gqlNode, gqlEdges);
   const isNew = gqlDiagram.layoutData.nodeLayoutData.find((nodeLayoutData) => nodeLayoutData.id === id) === undefined;
 
-  const data: FreeFormNodeData = {
+  const data: NodeData = {
     targetObjectId,
     targetObjectLabel,
     targetObjectKind,
     descriptionId,
     style: {
-      display: 'flex',
       backgroundColor: style.color,
-      borderColor: style.borderColor,
+      borderTopColor: style.borderColor,
+      borderBottomColor: style.borderColor,
+      borderLeftColor: style.borderColor,
+      borderRightColor: style.borderColor,
       borderRadius: style.borderRadius,
-      borderWidth: style.borderSize,
+      borderTopWidth: style.borderSize,
+      borderBottomWidth: style.borderSize,
+      borderLeftWidth: style.borderSize,
+      borderRightWidth: style.borderSize,
       borderStyle: convertLineStyle(style.borderStyle),
     },
     insideLabel: null,
@@ -76,7 +80,7 @@ const toRectangularNode = (
     defaultWidth: gqlNode.defaultWidth,
     defaultHeight: gqlNode.defaultHeight,
     isBorderNode: isBorderNode,
-    borderNodePosition: isBorderNode ? BorderNodePosition.EAST : null,
+    borderNodePosition: isBorderNode ? BorderNodePosition.WEST : null,
     labelEditable,
     positionDependentRotation: false,
     connectionHandles,
@@ -88,6 +92,7 @@ const toRectangularNode = (
     data.insideLabel = {
       id: insideLabel.id,
       text: insideLabel.text,
+      iconURL: labelStyle.iconURL,
       isHeader: insideLabel.isHeader,
       displayHeaderSeparator: insideLabel.displayHeaderSeparator,
       style: {
@@ -99,7 +104,6 @@ const toRectangularNode = (
         textAlign: 'center',
         ...convertLabelStyle(labelStyle),
       },
-      iconURL: labelStyle.iconURL,
     };
 
     const alignement = AlignmentMap[insideLabel.insideLabelLocation];
@@ -117,9 +121,9 @@ const toRectangularNode = (
     }
   }
 
-  const node: Node<FreeFormNodeData> = {
+  const node: Node<NodeData> = {
     id,
-    type: 'freeFormNode',
+    type: 'listNode',
     data,
     position: defaultPosition,
     hidden: state === GQLViewModifier.Hidden,
@@ -148,9 +152,32 @@ const toRectangularNode = (
   return node;
 };
 
-export class RectangleNodeConverterHandler implements INodeConverterHandler {
+const adaptChildrenBorderNodes = (nodes: Node[], gqlChildrenNodes: GQLNode<GQLNodeStyle>[]): void => {
+  const childrenNodes = nodes.filter(
+    (child) =>
+      gqlChildrenNodes.map((gqlChild) => gqlChild.id).find((gqlChildId) => gqlChildId === child.id) !== undefined
+  );
+  childrenNodes.forEach((child, index) => {
+    // Hide children node borders to prevent a 'bold' aspect.
+    child.data.style = {
+      ...child.data.style,
+      borderTopColor: 'transparent',
+      borderLeftColor: 'transparent',
+      borderRightColor: 'transparent',
+    };
+
+    if (index === childrenNodes.length - 1) {
+      child.data.style = {
+        ...child.data.style,
+        borderBottomColor: 'transparent',
+      };
+    }
+  });
+};
+
+export class ListNodeConverter implements INodeConverter {
   canHandle(gqlNode: GQLNode<GQLNodeStyle>) {
-    return gqlNode.style.__typename === 'RectangularNodeStyle' && gqlNode.childrenLayoutStrategy?.kind !== 'List';
+    return gqlNode.style.__typename === 'RectangularNodeStyle' && gqlNode.childrenLayoutStrategy?.kind === 'List';
   }
 
   handle(
@@ -164,7 +191,9 @@ export class RectangleNodeConverterHandler implements INodeConverterHandler {
     nodeDescriptions: GQLNodeDescription[]
   ) {
     const nodeDescription = nodeDescriptions.find((description) => description.id === gqlNode.descriptionId);
-    nodes.push(toRectangularNode(gqlDiagram, gqlNode, parentNode, nodeDescription, isBorderNode, gqlEdges));
+    if (nodeDescription) {
+      nodes.push(toListNode(gqlDiagram, gqlNode, parentNode, nodeDescription, isBorderNode, gqlEdges));
+    }
     convertEngine.convertNodes(
       gqlDiagram,
       gqlNode.borderNodes ?? [],
@@ -179,5 +208,6 @@ export class RectangleNodeConverterHandler implements INodeConverterHandler {
       nodes,
       nodeDescription?.childNodeDescriptions ?? []
     );
+    adaptChildrenBorderNodes(nodes, gqlNode.childNodes ?? []);
   }
 }
