+++ packages/diagrams/frontend/sirius-components-diagrams-reactflow/src/renderer/DiagramRenderer.tsx
@@ -84,17 +84,13 @@ export const DiagramRenderer = ({ diagramRefreshedEventPayload }: DiagramRendere
   const [nodes, setNodes, onNodesChange] = useNodesState<NodeData>([]);
   const [edges, setEdges, onEdgesChange] = useEdgesState<MultiLabelEdgeData>([]);
 
-  const { nodeConverterHandlers } = useContext<NodeTypeContextValue>(NodeTypeContext);
+  const { nodeConverters } = useContext<NodeTypeContextValue>(NodeTypeContext);
   const { fitToScreen } = useInitialFitToScreen();
   const { setSelection } = useSelection();
 
   useEffect(() => {
     const { diagram, cause } = diagramRefreshedEventPayload;
-    const convertedDiagram: Diagram = convertDiagram(
-      diagram,
-      nodeConverterHandlers,
-      diagramDescription.nodeDescriptions
-    );
+    const convertedDiagram: Diagram = convertDiagram(diagram, nodeConverters, diagramDescription.nodeDescriptions);
 
     if (cause === 'layout') {
       const selectedNodeIds = nodes.filter((node) => node.selected).map((node) => node.id);
